+++ Sources/WalletConnectRelay/ClientAuth/JWT/JWT.swift
@@ -4,6 +4,7 @@ struct JWT: Codable, Equatable {
     enum Errors: Error {
         case jwtNotSigned
     }
+    
     var header: Header
     var claims: Claims
     var signature: String? = nil
@@ -25,7 +26,6 @@ struct JWT: Codable, Equatable {
         let headerString = try header.encode()
         let claimsString = try claims.encode()
         return [headerString, claimsString, signature].joined(separator: ".")
-
     }
 }
 
@@ -56,16 +56,3 @@ extension JWT {
     }
 }
 
-
-struct JWTEncoder {
-    /// Returns a `String` representation of this data, encoded in base64url format
-    /// as defined in RFC4648 (https://tools.ietf.org/html/rfc4648).
-    ///
-    /// This is the appropriate format for encoding the header and claims of a JWT.
-    public static func base64urlEncodedString(data: Data) -> String {
-        let result = data.base64EncodedString()
-        return result.replacingOccurrences(of: "+", with: "-")
-            .replacingOccurrences(of: "/", with: "_")
-            .replacingOccurrences(of: "=", with: "")
-    }
-}
