+++ Sources/WalletConnectRelay/ClientAuth/JWT/JWTEncoder.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct JWTEncoder {
+    /// Returns a `String` representation of this data, encoded in base64url format
+    /// as defined in RFC4648 (https://tools.ietf.org/html/rfc4648).
+    ///
+    /// This is the appropriate format for encoding the header and claims of a JWT.
+    public static func base64urlEncodedString(data: Data) -> String {
+        let result = data.base64EncodedString()
+        return result.replacingOccurrences(of: "+", with: "-")
+            .replacingOccurrences(of: "/", with: "_")
+            .replacingOccurrences(of: "=", with: "")
+    }
+}
